requirements.txt
Flask==2.0.1
gunicorn==20.1.0
requests==2.21.0
jsonschema==4.2.1
beautifulsoup4==4.10.0
selenium==3.141.0
Pillow==8.3.2
urllib3==1.24.3
chromedriver-autoinstaller==0.3.1